+++ src/Plugins/LazyOptionalPlugin.php
@@ -322,6 +322,7 @@ class LazyOptionalPlugin implements Transformer
         foreach ($property->attributes as $attribute) {
             if (
                 $attribute->name === 'Spatie\\LaravelData\\Attributes\\Optional'
+                || $attribute->name === 'EffectSchemaGenerator\\Attributes\\Optional'
             ) {
                 return true;
             }